+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ### Added
 - utomatic error report via the internet using formspree.io API. 
+- Clean fallback methods of reporting errors like opening user's default email client with already filled out data, user just needs to hit send or opening a GitHub issue creation. 
 
 ### Changed
+- `global_exception_handler` to not create an archive with an error report and the associated log, but instead just an error report with the log pasted in. 
+- Automatic error reporting to be cleaner and handle any foreseeable problems that I saw. 
 
 ### Fixed